requirements.txt
fastapi
uvicorn
databases[postgresql]
sqlalchemy
python-dotenv
psycopg2-binary
pyjwt
passlib[bcrypt]